+++ javascript/libjoynr-js/src/main/js/joynr/dispatching/subscription/PublicationManager.js
@@ -792,6 +792,17 @@ define(
                     return false;
                 };
 
+                /* the parameter "callbackDispatcher" is optional, as in case of restoring
+                 * subscriptions, no reply must be sent back via the dispatcher
+                 */
+                function callbackDispatcherAsync(reply, callbackDispatcher) {
+                    if (callbackDispatcher !== undefined) {
+                        LongTimer.setTimeout(function asyncCallbackDispatcher() {
+                            callbackDispatcher(new SubscriptionReply(reply));
+                        }, 0);
+                    }
+                }
+
                 /**
                  * Handles SubscriptionRequests
                  *
@@ -859,12 +870,12 @@ define(
                                     subscriptionId : subscriptionId
                                 });
                                 log.error(exception.detailMessage);
-                                LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                    callbackDispatcher(new SubscriptionReply({
-                                        error : exception,
-                                        subscriptionId : subscriptionId
-                                    }));
-                                }, 0);
+                                callbackDispatcherAsync(
+                                        {
+                                            error : exception,
+                                            subscriptionId : subscriptionId
+                                        },
+                                        callbackDispatcher);
                                 return;
                             }
 
@@ -882,12 +893,12 @@ define(
                                     subscriptionId : subscriptionId
                                 });
                                 log.error(exception.detailMessage);
-                                LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                    callbackDispatcher(new SubscriptionReply({
-                                        error : exception,
-                                        subscriptionId : subscriptionId
-                                    }));
-                                }, 0);
+                                callbackDispatcherAsync(
+                                        {
+                                            error : exception,
+                                            subscriptionId : subscriptionId
+                                        },
+                                        callbackDispatcher);
                                 return;
                             }
 
@@ -906,12 +917,12 @@ define(
                                     subscriptionId : subscriptionId
                                 });
                                 log.error(exception.detailMessage);
-                                LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                    callbackDispatcher(new SubscriptionReply({
-                                        error : exception,
-                                        subscriptionId : subscriptionId
-                                    }));
-                                }, 0);
+                                callbackDispatcherAsync(
+                                        {
+                                            error : exception,
+                                            subscriptionId : subscriptionId
+                                        },
+                                        callbackDispatcher);
                                 return;
                             }
 
@@ -936,12 +947,12 @@ define(
                                         subscriptionId : subscriptionId
                                     });
                                     log.error(exception.detailMessage);
-                                    LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                        callbackDispatcher(new SubscriptionReply({
-                                            error : exception,
-                                            subscriptionId : subscriptionId
-                                        }));
-                                    }, 0);
+                                    callbackDispatcherAsync(
+                                            {
+                                                error : exception,
+                                                subscriptionId : subscriptionId
+                                            },
+                                            callbackDispatcher);
                                     return;
                                 }
 
@@ -968,12 +979,12 @@ define(
                                         subscriptionId : subscriptionId
                                     });
                                     log.error(exception.detailMessage);
-                                    LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                        callbackDispatcher(new SubscriptionReply({
-                                            error : exception,
-                                            subscriptionId : subscriptionId
-                                        }));
-                                    }, 0);
+                                    callbackDispatcherAsync(
+                                            {
+                                                error : exception,
+                                                subscriptionId : subscriptionId
+                                            },
+                                            callbackDispatcher);
                                     return;
                                 }
                                 // call the get method on the provider at the set interval
@@ -1001,11 +1012,11 @@ define(
                                     sendPublication(subscriptionInfo, undefined, exception);
                                     return exception;
                                 });
-                            LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                callbackDispatcher(new SubscriptionReply({
-                                    subscriptionId : subscriptionId
-                                }));
-                            }, 0);
+                            callbackDispatcherAsync(
+                                    {
+                                        subscriptionId : subscriptionId
+                                    },
+                                    callbackDispatcher);
                         };
 
                 /**
@@ -1075,12 +1086,12 @@ define(
                                     subscriptionId : subscriptionId
                                 });
                                 log.error(exception.detailMessage);
-                                LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                    callbackDispatcher(new SubscriptionReply({
-                                        error : exception,
-                                        subscriptionId : subscriptionId
-                                    }));
-                                }, 0);
+                                callbackDispatcherAsync(
+                                        {
+                                            error : exception,
+                                            subscriptionId : subscriptionId
+                                        },
+                                        callbackDispatcher);
                                 return;
                             }
 
@@ -1108,12 +1119,12 @@ define(
                                     subscriptionId : subscriptionId
                                 });
                                 log.error(exception.detailMessage);
-                                LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                    callbackDispatcher(new SubscriptionReply({
-                                        error : exception,
-                                        subscriptionId : subscriptionId
-                                    }));
-                                }, 0);
+                                callbackDispatcherAsync(
+                                        {
+                                            error : exception,
+                                            subscriptionId : subscriptionId
+                                        },
+                                        callbackDispatcher);
                                 return;
                             }
 
@@ -1138,12 +1149,12 @@ define(
                                         subscriptionId : subscriptionId
                                     });
                                     log.error(exception.detailMessage);
-                                    LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                        callbackDispatcher(new SubscriptionReply({
-                                            error : exception,
-                                            subscriptionId : subscriptionId
-                                        }));
-                                    }, 0);
+                                    callbackDispatcherAsync(
+                                            {
+                                                error : exception,
+                                                subscriptionId : subscriptionId
+                                            },
+                                            callbackDispatcher);
                                     return;
                                 }
 
@@ -1161,11 +1172,11 @@ define(
 
                             persistency.setItem(subscriptionId, JSON.stringify(subscriptionInfo));
                             storeSubscriptions();
-                            LongTimer.setTimeout(function asyncCallbackDispatcher() {
-                                callbackDispatcher(new SubscriptionReply({
-                                    subscriptionId : subscriptionId
-                                }));
-                            }, 0);
+                            callbackDispatcherAsync(
+                                    {
+                                        subscriptionId : subscriptionId
+                                    },
+                                    callbackDispatcher);
                         };
 
                 /**